wca_nr_api/requirements.txt
uvicorn
fastapi
pandas
pydantic
gcsfs
fsspec
google-cloud-storage
